requirements.txt
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai
Flask==2.3.3
Flask-CORS==4.0.0